requirements.txt
pandas
smart-open==1.8.0
pyarrow
python-dotenv